refactor: create callbacks directory
